Updated Stable Matter description
Stable Matter had a cool description, but for a nerd like me, there were some issues I noticed. The first, is that Stable Matter does not adhere to physics, which is as absurd as saying "an element that can't be put on the periodic table": It doesn't make any sense. So, I changed it to read that it does not adhere to physics as we know them, which acknowledges that it definitely still follows physics, just none that we currently understand.

I also tweaked the explanation as to why it works. It also didn't make much sense in terms of realism.

This description is definitely not complete (It's really long), but it should be enough to make the description make more sense.